move tests from snapshot to manifest list reader
